try fixing delete role

+++ .evergreen/csfle/azurekms/delete-vm.sh
@@ -12,14 +12,17 @@ if [ -z "${AZUREKMS_RESOURCEGROUP:-}" ] || \
     exit 1
 fi
 
-echo "Deleting the role from the Virtual Machine $AZUREKMS_VMNAME ... begin"
-PRINCIPAL_ID=$(az vm show --show-details --resource-group "$AZUREKMS_RESOURCEGROUP" --name "$AZUREKMS_VMNAME" --query identity.principalId -o tsv)
-az role assignment delete \
-    --assignee-object-id "$PRINCIPAL_ID" \
-    --role "Key Vault Crypto User" \
-    -y \
-    >/dev/null
-echo "Deleting the role from the Virtual Machine $AZUREKMS_VMNAME ... end"
+if [ -n "${$AZUREKMS_SCOPE}" ]; then
+    echo "Deleting the role from the Virtual Machine $AZUREKMS_VMNAME ... begin"
+    PRINCIPAL_ID=$(az vm show --show-details --resource-group "$AZUREKMS_RESOURCEGROUP" --name "$AZUREKMS_VMNAME" --query identity.principalId -o tsv)
+    az role assignment delete \
+        --assignee "$PRINCIPAL_ID" \
+        --role "Key Vault Crypto User" \
+        --scope "$AZUREKMS_SCOPE"
+        -y \
+        >/dev/null
+    echo "Deleting the role from the Virtual Machine $AZUREKMS_VMNAME ... end"
+fi
 
 echo "Deleting Virtual Machine $AZUREKMS_VMNAME ... begin"
 az vm delete \
